Add vector_math Cython kernels and clean up ops
- Implement CEILING/FLOOR/TRUNCATE/POWER/RANDOM/RANDOM_NORMAL in new
  opteryx/compiled/vector_ops/vector_math.pyx; add a seeded RNG for
  reproducible normal draws
- Include vector_math in the vector_ops build
- Add StringConcat fast path in evaluator to keep inputs as Draken
  strings/bytes and avoid Arrow conversion
- Change expression literal FLOAT mapping to OrsoTypes.DOUBLE
- Remove legacy opteryx/expression/ops.py and strip pyarrow/numpy
  imports from several modules

+++ opteryx/expression/__init__.py
@@ -463,7 +463,7 @@ def _inner_evaluate(root: Node, table):
         if literal_type == OrsoTypes.INTEGER:
             return Int64Vector.from_constant(int(value), length)
 
-        if literal_type == OrsoTypes.FLOAT:
+        if literal_type == OrsoTypes.DOUBLE:
             return Float64Vector.from_constant(float(value), length)
 
         if literal_type == OrsoTypes.BOOLEAN:
